% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jags_combine.R
\name{jags_combine}
\alias{jags_combine}
\title{Combine individual MCMC chains into an mcmc.list object}
\usage{
jags_combine(...)
}
\arguments{
\item{\dots}{mcmc object}
}
\value{
An \code{mcmc.list}
}
\description{
\code{jags_combine} is a helper function for running jags in parallel
     that combines the individual chains run on separate cores into an
     mcmc.list object. This function is supplied to the .combine argument in
     the foreach function.
}
\examples{
\dontrun{jags_combine(result1, result2, result3)}

}
\seealso{
\code{\link{mcmc}} \code{\link{mcmc.list}}
}
\author{
Michael Malick
}
